Fix CacheHead dtype mismatch under --precision bf16
TimestepAdaLN.forward fed the sinusoidal timestep embedding straight into its
projection. sinusoidal_embedding_1d always returns float32 (it builds the
frequency table with dtype=torch.float32 and calls timestep.float()), while the
projection weights follow the model dtype. Under --precision bf16 that made
F.linear raise "mat1 and mat2 must have the same dtype, but got Float and
BFloat16" on the very first head call.

Wan's own DiT casts at exactly this point --
self.time_embedding(sinusoidal_embedding_1d(...).to(x.dtype)) -- CacheHead did
not. The cast here targets the projection weight dtype, which is what F.linear
actually requires, rather than the activation dtype.

Every existing test ran in float32, where the mismatch cannot occur, so the
whole suite passed while any real bf16 run failed immediately. Adds a
precision matrix over float32/bfloat16/float16 covering forward, backward, and
a float32 timestep against a bf16 model (the scheduler builds timesteps in
float32 regardless of --precision).

Also accumulates the supervised per-step loss in float32: huber_loss returns
bf16 under bf16, and summing ten of those then dividing discards precision in
both the logged value and the gradient.

+++ examples/wanvideo/model_inference/cache_head/cache_head_model_training.py
@@ -483,7 +483,11 @@ class CacheHeadTrainer:
                         f"{expected_tokens + (WAN_NOISE_TOKEN_CHANNELS,)}, got "
                         f"head {tuple(v_tokens.shape)} vs teacher {tuple(teacher_tokens.shape)}"
                     )
-                step_loss = regression_loss(v_tokens, teacher_tokens, self.reg_loss)
+                # Accumulate in float32: under --precision bf16 the loss comes
+                # back as bf16 (~3 decimal digits), and summing ten of them then
+                # dividing throws away precision in both the logged value and
+                # the gradient.  Autograd carries the cast back through.
+                step_loss = regression_loss(v_tokens, teacher_tokens, self.reg_loss).float()
                 total = step_loss if total is None else total + step_loss
                 per_step.append({"step": k, "loss": float(step_loss.detach().item())})
                 noise_pred = unpatchify_tokens(v_tokens, self.grid, self.patch)
